test: refactor test imports
