+++ tests/integration/mcp-proxy-mock.test.ts
@@ -10,8 +10,8 @@ import {
 	type ServerResponse,
 } from "node:http";
 import { afterEach, beforeEach, describe, test } from "node:test";
-import type { Logger } from "../../src/logger.ts";
-import { createCamundaFetch } from "../../src/mcp-proxy-helpers.ts";
+import type { Logger } from "../../src/core/logger.ts";
+import { createCamundaFetch } from "../../src/utils/mcp-proxy-helpers.ts";
 import { makeMockClient, makeMockLogger } from "../utils/mocks.ts";
 
 describe("MCP Proxy Mock Server Integration Tests", () => {
