+++ tests/unit/command-dispatch-structure.test.ts
@@ -17,11 +17,11 @@
 import assert from "node:assert";
 import { describe, test } from "node:test";
 
-import { COMMAND_DISPATCH } from "../../src/command-dispatch.ts";
+import { COMMAND_DISPATCH } from "../../src/framework/command-dispatch.ts";
 import {
 	DEFINE_COMMAND_MARKER,
 	isDefinedCommand,
-} from "../../src/command-framework.ts";
+} from "../../src/framework/command-framework.ts";
 
 describe("COMMAND_DISPATCH structural invariant (#290)", () => {
 	test("every dispatch entry is produced by defineCommand()", () => {
